requirements.txt
pydantic-settings~=2.8.1
fastapi_cache2~=0.2.2
python-multipart~=0.0.20
pydantic~=2.10.5
sqlmodel~=0.0.23
pydantic_core~=2.27.2
PyJWT~=2.10.1
passlib~=1.7.4
fastapi~=0.115.8
uvicorn~=0.34.0
httpx~=0.28.1
SQLAlchemy~=2.0.38
asyncio~=3.4.3
slowapi~=0.1.9
alembic~=1.15.1
APScheduler~=3.11.0
redis~=4.6.0
# sudo apt install python3-dev libpq-dev
psycopg2~=2.9.10